+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.52.1] - 2025-10-04
+
+### Changed
+
+- Added Cargo examples
+- aws-sdk-s3 = "1.107.0"
+- Updated dependencies
+- Updated docs
+
+### Fixed
+
+- Fixed a bug that `Pipeline::get_errors_and_consume()` does not return errors in certain cases.
 
 ## [1.52.0] - 2025-10-02